medical_retrieval/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
pillow>=10.0.0

# Medical CLIP models
open_clip_torch>=2.20.0
huggingface-hub>=0.19.0

# Vector search
faiss-gpu>=1.7.2  # Use faiss-cpu if no GPU
numpy>=1.24.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
tqdm>=4.66.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Monitoring
wandb>=0.16.0  # Optional: for experiment tracking
